clean up compiler warning in envenom
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@1308 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ sc_rogue.cpp
@@ -1326,7 +1326,7 @@ struct envenom_t : public rogue_attack_t
 
     rogue_t* p = player -> cast_rogue();
 
-    double doses_consumed = std::min( p -> buffs_poison_doses, p -> buffs_combo_points );
+    int doses_consumed = std::min( p -> buffs_poison_doses, p -> buffs_combo_points );
 
     double envenom_duration = 1.0 + p -> buffs_combo_points;
